+++ test/support/js/full-app/index.js
@@ -32,4 +32,9 @@ app.get('/static').to(ctx => ctx.sendFile(ctx.home.child('public', 'test.txt')))
 
 app.get('/default/view').to('foo#defaultView');
 
+app.websocket('/echo.json').to('foo#websocket');
+
+app.get('/session/login/:name').to('auth#login');
+app.get('/session/logout').to('auth#logout');
+
 app.start();